python/pyproject.toml
[tool.poetry]
name = "telegram-scrapper"
version = "0.1"
description = ""
authors = ["hegemonies <hegemonies@yandex.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.19.2"
Telethon = "^1.24.0"
cryptg = "^0.2.post4"
requests = "^2.26.0"
schedule = "^1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
